requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
decorator==5.1.1
Django==4.1.1
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dynaconf==3.1.9
hvac==0.11.2
idna==3.3
knox==0.1.14
loguru==0.6.0
PyJWT==2.4.0
pytz==2022.2.1
requests==2.28.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
validators==0.20.0
win32-setctime==1.1.0
